Additional work and refactoring on graph rendering. The first work on categories was done.

+++ rcp/com.puresol.rcp.coding.client/com.puresol.rcp.coding.client.plugin/src/com/puresol/coding/client/charting/rendering/TransformationMatrix2D.java
@@ -59,7 +59,7 @@ public class TransformationMatrix2D extends Matrix {
 	}
     }
 
-    public void translate(int x, int y) {
+    public void translate(double x, double y) {
 	for (int i = 0; i < DIMENSIONS; i++) {
 	    set(i, 2, get(i, 0) * x + get(i, 1) * y + get(i, 2));
 	}
